+++ tests/orbit_arcball_behavior_test.cpp
@@ -3,7 +3,7 @@
  * Licensed under the Apache License, Version 2.0 (the "License")
  * --------------------------------------------------------------------- */
 
-#include "vertexnova/interaction/orbit_behavior.h"
+#include "vertexnova/interaction/orbit_arcball_behavior.h"
 #include "vertexnova/scene/camera/camera_factory.h"
 #include "vertexnova/scene/camera/camera_types.h"
 
@@ -17,38 +17,38 @@ static std::shared_ptr<vne::scene::PerspectiveCamera> makePerspCamera() {
         vne::scene::PerspectiveCameraParameters(45.0f, 16.0f / 9.0f, 0.1f, 1000.0f));
 }
 
-TEST(OrbitBehavior, DefaultValues) {
-    vne::interaction::OrbitBehavior b;
+TEST(OrbitArcballBehavior, DefaultValues) {
+    vne::interaction::OrbitArcballBehavior b;
     EXPECT_EQ(b.getRotationMode(), vne::interaction::OrbitRotationMode::eEuler);
     EXPECT_EQ(b.getPivotMode(), vne::interaction::OrbitPivotMode::eCoi);
     EXPECT_GT(b.getOrbitDistance(), 0.0f);
     EXPECT_GT(b.getZoomSpeed(), 0.0f);
 }
 
-TEST(OrbitBehavior, SetRotationMode) {
-    vne::interaction::OrbitBehavior b;
+TEST(OrbitArcballBehavior, SetRotationMode) {
+    vne::interaction::OrbitArcballBehavior b;
     b.setRotationMode(vne::interaction::OrbitRotationMode::eQuaternion);
     EXPECT_EQ(b.getRotationMode(), vne::interaction::OrbitRotationMode::eQuaternion);
 }
 
-TEST(OrbitBehavior, SetPivotMode) {
-    vne::interaction::OrbitBehavior b;
+TEST(OrbitArcballBehavior, SetPivotMode) {
+    vne::interaction::OrbitArcballBehavior b;
     b.setPivotMode(vne::interaction::OrbitPivotMode::eFixed);
     EXPECT_EQ(b.getPivotMode(), vne::interaction::OrbitPivotMode::eFixed);
 }
 
-TEST(OrbitBehavior, SetOrbitDistanceClamped) {
-    vne::interaction::OrbitBehavior b;
+TEST(OrbitArcballBehavior, SetOrbitDistanceClamped) {
+    vne::interaction::OrbitArcballBehavior b;
     b.setOrbitDistance(0.001f);
     EXPECT_GE(b.getOrbitDistance(), 0.01f);
 }
 
-TEST(OrbitBehavior, CameraIntegration) {
+TEST(OrbitArcballBehavior, CameraIntegration) {
     auto cam = makePerspCamera();
     cam->setPosition(vne::math::Vec3f(0.0f, 0.0f, 5.0f));
     cam->lookAt(vne::math::Vec3f(0.0f, 0.0f, 0.0f), vne::math::Vec3f(0.0f, 1.0f, 0.0f));
 
-    vne::interaction::OrbitBehavior b;
+    vne::interaction::OrbitArcballBehavior b;
     b.setCamera(cam);
     b.setViewportSize(1280.0f, 720.0f);
 
@@ -66,9 +66,9 @@ TEST(OrbitBehavior, CameraIntegration) {
     EXPECT_GT((cam->getPosition() - vne::math::Vec3f(0.0f, 0.0f, 5.0f)).length(), 0.01f);
 }
 
-TEST(OrbitBehavior, FitToAABB) {
+TEST(OrbitArcballBehavior, FitToAABB) {
     auto cam = makePerspCamera();
-    vne::interaction::OrbitBehavior b;
+    vne::interaction::OrbitArcballBehavior b;
     b.setCamera(cam);
     b.setViewportSize(1280.0f, 720.0f);
 
@@ -80,8 +80,8 @@ TEST(OrbitBehavior, FitToAABB) {
     EXPECT_NEAR(coi.z(), 0.0f, 1e-3f);
 }
 
-TEST(OrbitBehavior, ResetState) {
-    vne::interaction::OrbitBehavior b;
+TEST(OrbitArcballBehavior, ResetState) {
+    vne::interaction::OrbitArcballBehavior b;
     EXPECT_NO_FATAL_FAILURE(b.resetState());
 }